feat: add auth middleware

+++ models/event.go
@@ -12,7 +12,7 @@ type Event struct {
 	Description string    `binding: "required"`
 	Location    string    `binding: "required"`
 	DateTime    time.Time `binding: "required"`
-	UserID      int       `binding: "required"`
+	UserID      int64     `binding: "required"`
 }
 
 var events = []Event{}
